migrate to MPS 2.5

+++ languages/query/source_gen/jetbrains/mps/core/query/runtime/CompositeScope.java
@@ -45,7 +45,7 @@ public class CompositeScope extends QueryScope {
         ListSequence.fromList(result).addElement(r);
       }
     }
-    return (ListSequence.fromList(result).count() == 1 ?
+    return ((int) ListSequence.fromList(result).count() == 1 ?
       ListSequence.fromList(result).first() :
       null
     );
